Style and custom configuration settings
Updated the README to indicate work in progress.
Updated CSS and references to the correct CSS file.
Updated the compose.json file to name the project and version.
Added configuration setting to the config-dist.
Removed a bunch of unnecessary session variables from SLAM.php and added SLAM-specific wording where appropriate.